requirements.txt
aplpy==2.1.0
astropy==5.0.4
astropy-healpix==0.6
attrs==21.4.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
codecov==2.1.13
confluent-kafka==2.2
coverage==6.3.2
coveralls==3.3.1
cycler==0.11.0
Cython==0.29.28
distlib==0.3.4
docopt==0.6.2
fastavro==1.4.10
filelock==3.6.0
fink-filters==2.2
fink-utils==0.4.0
fink-client==2.11
fonttools==4.31.2
identify==2.4.12
idna==3.3
imageio==2.16.1
iniconfig==1.1.1
kiwisolver==1.4.2
matplotlib==3.5.1
networkx==2.7.1
nodeenv==1.6.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.18.1
py==1.11.0
PyAVM==0.9.5
pyerfa==2.0.0.1
pyparsing==3.0.7
pyregion==2.1.1
pytest==7.1.1
python-dateutil==2.8.2
python-snappy==0.6.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
reproject==0.8
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
Shapely==1.8.1.post1
six==1.16.0
tabulate==0.8.9
tifffile==2022.3.25
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
virtualenv==20.14.1
visdcc==0.0.50
pyspark==3.2.2
pyarrow==7.0.0
